Moved security filters into package
