--- src/main/java/uk/gov/companieshouse/account/validator/SecurityConstants.java
@@ -1,7 +1,0 @@
-package uk.gov.companieshouse.account.validator;
-
-public class SecurityConstants {
-
-    public static final String ACTUATOR_URL = "/actuator/**";
-
-}
