+++ src/test/java/uk/gov/companieshouse/account/validator/AuthenticationFilterTest.java
@@ -1,18 +1,20 @@
 package uk.gov.companieshouse.account.validator;
 
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.test.context.SpringBootTest;
+import uk.gov.companieshouse.account.validator.security.AuthenticationFilter;
 import uk.gov.companieshouse.logging.Logger;
 
 import javax.servlet.FilterChain;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class AuthenticationFilterTest {
